renaming linting -> lint, moved PR template, template_struct moved
