--- .github/PULL_REQUEST_TEMPLATE/pull_request.md
@@ -1,32 +1,0 @@
----
-name: Pull request
-about: Submit a pull request to COOKIETEMPLE
-title: '[TEMPLATENAME/COMMAND/CORE] Pull Request Title'
-labels: ''
-assignees: ''
-
----
-
-Many thanks for contributing to COOKIETEMPLE!
-
-**Associated Template/Command/Core**
-State the template handle (e.g. python-cli) and also add it in the title in the square brackets.
-If your pull request is not directly related to any template, please state either the associated COOKIETEMPLE command or '[CORE]' if it is a general COOKIETEMPLE bug.
-
-**PR Checklist**
-Please fill in the appropriate checklist below (delete whatever is not relevant). These are the most common things requested on pull requests (PRs).
-
- - [ ] This comment contains a description of changes (with reason)
- - [ ] Referenced issue is linked
- - [ ] If you've fixed a bug or added code that should be tested, add tests!
- - [ ] Documentation in `docs` is updated
- - [ ] `CHANGELOG.rst` is updated
-
-**Description of changes**
-Please state what you've changed and how it might affect the user.
-
-**Technical details**
-Please state any technical details such as limitations, reasons for additional dependencies, benchmarks etc. here.
-
-**Additional context**
-Add any other context or screenshots here.
