+++ cookietemple/create/domains/PubCreator.py
@@ -5,7 +5,7 @@ import click
 
 from cookietemple.create.TemplateCreator import TemplateCreator
 from cookietemple.create.github_support import load_github_username
-from cookietemple.util.cookietemple_template_struct import CookietempleTemplateStruct
+from cookietemple.create.domains.cookietemple_template_struct import CookietempleTemplateStruct
 
 
 @dataclass
