+++ cookietemple/lint/domains/web.py
@@ -3,12 +3,12 @@ from subprocess import Popen
 
 import click
 
-from cookietemple.linting.TemplateLinter import TemplateLinter, files_exist_linting
+from cookietemple.lint.TemplateLinter import TemplateLinter, files_exist_linting
 
 CWD = os.getcwd()
 
 
-class CliPythonLint(TemplateLinter):
+class WebWebsitePythonLint(TemplateLinter):
     def __init__(self, path):
         super().__init__(path)
 
@@ -32,7 +32,6 @@ class CliPythonLint(TemplateLinter):
             'tox.ini',
         Files that *should* be present::
             '.github/workflows/build_package.yml',
-            '.github/workflows/publish_package.yml',
             '.github/workflows/tox_testsuite.yml',
             '.github/workflows/flake8.yml',
         Files that *must not* be present::
@@ -51,7 +50,6 @@ class CliPythonLint(TemplateLinter):
         ]
         files_warn = [
             [os.path.join('.github', 'workflows', 'build_package.yml')],
-            [os.path.join('.github', 'workflows', 'publish_package.yml')],
             [os.path.join('.github', 'workflows', 'tox_testsuite.yml')],
             [os.path.join('.github', 'workflows', 'flake8_linting.yml')],
         ]
