+++ cookietemple/util/click_util.py
@@ -16,7 +16,7 @@ class CustomHelpOrder(click.Group):
         return super(CustomHelpOrder, self).get_help(ctx)
 
     def list_commands_for_help(self, ctx):
-        """reorder the list of commands when listing the help"""
+        '""reorder the list of commands when listing the help""'
         commands = super(CustomHelpOrder, self).list_commands(ctx)
         return (c[1] for c in sorted((self.help_priorities.get(command, 1000), command) for command in commands))
